Added accuracy check & fixed negative damage oversight

+++ PMDMMO-common/src/com/darkxell/common/pokemon/AffectsPokemon.java
@@ -7,7 +7,17 @@ import com.darkxell.common.pokemon.BaseStats.Stat;
 /** Any game object that may modify stats or behavior of a Pokemon should implement this interface, for consistency.<br>
  * This interface has default methods and only methods that are necessary can be implemented.<br>
  * This allows for all methods that may affect a Pokemon's stats or behavior to be part of this interface. <br>
- * Some objects that implement this interfact do not <i>belong</i> to a Pokemon (such as Weather); in that case, <code>isUser</code> booleans in method arguments should be ignored. */
+ * Some objects that implement this interface do not <i>belong</i> to a Pokemon (such as Weather); in that case, <code>isUser</code> booleans in method arguments should be ignored.<br>
+ * <br>
+ * 
+ * This Interface allows the alteration of:<br>
+ * <ul>
+ * <li>Critical hit rates</li>
+ * <li>Stat stages</li>
+ * <li>Stat values</li>
+ * <li>Damage multiplier</li>
+ * </ul>
+ */
 public interface AffectsPokemon
 {
 
@@ -24,7 +34,7 @@ public interface AffectsPokemon
 		return critical;
 	}
 
-	/** Called when a Pokemon uses a damaging move. Modifies the value of the input Stat. (If attack, it's for the user; if defense, it's for the target.)
+	/** Called when a Pokemon uses a damaging move. Modifies the value of the input Stat. (If attack or accuracy, it's for the user; if defense or evasion, it's for the target.)
 	 * 
 	 * @param stat - The Stat to modify.
 	 * @param value - The value of the Stat before this object applies its modifications.
@@ -33,12 +43,12 @@ public interface AffectsPokemon
 	 * @param isUser - <code>true</code> if this Object belongs to the Move's user (if it's its ability or an item it holds).
 	 * 
 	 * @return The new value of the Stat with modifications applied by this object. */
-	public default int applyStatModifications(Stat stat, int value, MoveUse move, DungeonPokemon target, boolean isUser, Floor floor)
+	public default double applyStatModifications(Stat stat, double value, MoveUse move, DungeonPokemon target, boolean isUser, Floor floor)
 	{
 		return value;
 	}
 
-	/** Called when a Pokemon uses a damaging move. Modifies the stage of the input Stat. (If attack, it's for the user; if defense, it's for the target.)
+	/** Called when a Pokemon uses a damaging move. Modifies the stage of the input Stat. (If attack or accuracy, it's for the user; if defense or evasion, it's for the target.)
 	 * 
 	 * @param stat - The Stat to modify.
 	 * @param value - The value of the Stat before this object applies its modifications.
